ref: AbstractReport
